Major reorg of Aircloak namespace.
Reorganised namespace hierarchy.
Extracted individual classes into their own files.
Applied some code Analyzer suggestions (more to follow).
