+++ src/aircloak/Aircloak/JsonApi/JsonApiClient.cs
@@ -1,0 +1,147 @@
+﻿namespace Aircloak.JsonApi
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Net;
+    using System.Net.Http;
+    using System.Net.Http.Headers;
+    using System.Text.Json;
+    using System.Threading.Tasks;
+
+    using Aircloak.JsonApi.ResponseTypes;
+
+    /// <summary>
+    /// Convenience class derived from <c>HttpClient</c> provides GET and POST methods adapted to the
+    /// Aircloak API:
+    /// <list type="bullet">
+    /// <item>
+    /// <description>Sets the provided Api Key on all outgoing requests.</description>
+    /// </item>
+    /// <item>
+    /// <description>Augments unsuccessful requests with custom error messages. </description>
+    /// </item>
+    /// <item>
+    /// <description>Deserializes Json responses.</description>
+    /// </item>
+    /// </list>
+    /// </summary>
+    public class JsonApiClient : HttpClient
+    {
+        /// <summary>
+        /// Send a GET request to the Aircloak API. Handles authentication.
+        /// </summary>
+        /// <param name="apiEndpoint">The API endpoint to target.</param>
+        /// <param name="apiKey">The API key for the service.</param>
+        /// <typeparam name="T">The type to deserialize the JSON response to.</typeparam>
+        /// <returns>A <c>Task&lt;T&gt;</c> which, upon completion, contains the API response deserialized
+        /// to the provided return type.</returns>
+        /// <exception cref="HttpRequestException"></exception>
+        /// <exception cref="JsonException">The JSON is invalid.
+        /// -or- <c>T</c> is not compatible with the JSON.
+        /// -or- There is remaining data in the stream.</exception>
+        public async Task<T> ApiGetRequest<T>(
+            Uri apiEndpoint,
+            string apiKey)
+        {
+            return await ApiRequest<T>(HttpMethod.Get, apiEndpoint, apiKey);
+        }
+
+        /// <summary>
+        /// Send a POST request to the Aircloak API. Handles authentication.
+        /// </summary>
+        /// <param name="apiEndpoint">The API endpoint to target.</param>
+        /// <param name="apiKey">The API key for the service.</param>
+        /// <param name="requestContent">JSON-encoded request message (optional).</param>
+        /// <typeparam name="T">The type to deserialize the JSON response to.</typeparam>
+        /// <returns>A <c>Task&lt;T&gt;</c> which, upon completion, contains the API response deserialized
+        /// to the provided return type.</returns>
+        /// <exception cref="HttpRequestException"></exception>
+        /// <exception cref="JsonException">The JSON is invalid.
+        /// -or- <c>T</c> is not compatible with the JSON.
+        /// -or- There is remaining data in the stream.</exception>
+        public async Task<T> ApiPostRequest<T>(
+            Uri apiEndpoint,
+            string apiKey,
+            string requestContent = default)
+        {
+            return await ApiRequest<T>(HttpMethod.Post, apiEndpoint, apiKey, requestContent);
+        }
+
+        /// <summary>
+        /// Send a request to the Aircloak API. Handles authentication.
+        /// </summary>
+        /// <param name="requestMethod">The HTTP method to use in the request.</param>
+        /// <param name="apiEndpoint">The API endpoint to target.</param>
+        /// <param name="apiKey">The API key for the service.</param>
+        /// <param name="requestContent">JSON-encoded request message (optional).</param>
+        /// <typeparam name="T">The type to deserialize the JSON response to.</typeparam>
+        /// <returns>A <c>Task&lt;T&gt;</c> which, upon completion, contains the API response deserialized
+        /// to the provided return type.</returns>
+        /// <exception cref="HttpRequestException"></exception>
+        /// <exception cref="JsonException">The JSON is invalid.
+        /// -or- <c>T</c> is not compatible with the JSON.
+        /// -or- There is remaining data in the stream.</exception>
+        private async Task<T> ApiRequest<T>(
+            HttpMethod requestMethod,
+            Uri apiEndpoint,
+            string apiKey,
+            string requestContent = default)
+        {
+            using var requestMessage =
+                new HttpRequestMessage(requestMethod, apiEndpoint);
+
+            if (!requestMessage.Headers.TryAddWithoutValidation("auth-token", apiKey))
+            {
+                throw new Exception($"Failed to add Http header 'auth-token: {apiKey}'");
+            }
+
+            if (!(requestContent is null))
+            {
+                requestMessage.Content = new StringContent(requestContent);
+                requestMessage.Content.Headers.ContentType = new MediaTypeHeaderValue("application/json");
+            }
+
+            using var response = await SendAsync(
+                requestMessage,
+                HttpCompletionOption.ResponseHeadersRead);
+
+            if (response.IsSuccessStatusCode)
+            {
+                using var contentStream = await response.Content.ReadAsStreamAsync();
+                var opts = new JsonSerializerOptions
+                {
+                    PropertyNamingPolicy = new SnakeCaseNamingPolicy(),
+                };
+                return await JsonSerializer.DeserializeAsync<T>(contentStream, opts);
+            }
+            else
+            {
+                throw new HttpRequestException($"{requestMethod} Request Error: {ServiceError(response)}");
+            }
+        }
+
+        /// <summary>
+        /// Turns the HTTP response into a custom error string.
+        /// </summary>
+        /// <param name="response">The HTTP response.</param>
+        /// <returns>A string containing a custom error message.</returns>
+        private string ServiceError(HttpResponseMessage response)
+        {
+            return response.StatusCode switch
+            {
+                HttpStatusCode.Unauthorized =>
+                    "Unauthorized -- Your API token is wrong",
+                HttpStatusCode.NotFound =>
+                    "Not Found -- Invalid URL",
+                HttpStatusCode.InternalServerError =>
+                    "Internal Server Error -- We had a problem with our server. Try again later.",
+                HttpStatusCode.ServiceUnavailable =>
+                    "Service Unavailable -- We're temporarily offline for maintenance. Please try again later.",
+                HttpStatusCode.GatewayTimeout =>
+                    "Gateway Timeout -- A timeout occured while contacting the data source. " +
+                    "The system might be overloaded. Try again later.",
+                _ => response.StatusCode.ToString(),
+            };
+        }
+    }
+}
