+++ src/aircloak/Aircloak/JsonApi/SnakeCaseNamingPolicy.cs
@@ -1,12 +1,12 @@
-﻿using System;
-using System.Linq;
-using System.Text.Json;
-using System.Text.RegularExpressions;
-
-namespace Aircloak.SnakeCaseNamingPolicy
+﻿namespace Aircloak.JsonApi
 {
+    using System;
+    using System.Linq;
+    using System.Text.Json;
+    using System.Text.RegularExpressions;
+
     /// <summary>
-    /// Helper class for converting Json keys from the .NET standard PascalCase to snake_case
+    /// Helper class for converting Json keys from the .NET standard PascalCase to snake_case.
     /// </summary>
     public class SnakeCaseNamingPolicy : JsonNamingPolicy
     {
@@ -14,7 +14,7 @@ namespace Aircloak.SnakeCaseNamingPolicy
         {
             var fragments = Regex.Matches(pascalCase, "[A-Z]+[a-z]+")
                 .Select(match => match.Value.ToLower());
-            return String.Join("_", fragments);
+            return string.Join("_", fragments);
         }
     }
